requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==2.1.1
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
